+++ DenseNeuralNetwork/Loss.cpp
@@ -7,8 +7,8 @@ Loss* Loss::ACCURACY = { new Accuracy() };
 Loss* Loss::BINARY_ACCURACY = { new BinaryAccuracy() };
 Loss* Loss::ALL_LOSSES[Loss::NUM_LOSSES] = {MEAN_SQUARED_ERROR, BINARY_CROSS_ENTROPY, CATEGORICAL_CROSS_ENTROPY};
 
-double MeanSquaredError::loss(Layer* layer, double** yTrue) {
-	double sum = 0;
+float MeanSquaredError::loss(Layer* layer, float** yTrue) {
+	float sum = 0;
 	for (int i = 0; i < layer->batchSize; i++) {
 		for (int j = 0; j < layer->size; j++) {
 			sum += (layer->neurons[i][j] - yTrue[i][j]) * (layer->neurons[i][j] - yTrue[i][j]);
@@ -17,7 +17,7 @@ double MeanSquaredError::loss(Layer* layer, double** yTrue) {
 	return sum / layer->size;
 }
 
-void MeanSquaredError::differentiate(Layer* layer, double** yTrue) {
+void MeanSquaredError::differentiate(Layer* layer, float** yTrue) {
 	for (int i = 0; i < layer->batchSize; i++) {
 		for (int j = 0; j < layer->size; j++) {
 			layer->neuronGradient[i][j] = 2 * (layer->neurons[i][j] - yTrue[i][j]) / layer->size;
@@ -30,8 +30,8 @@ string MeanSquaredError::toString() {
 	return toString;
 }
 
-double BinaryCrossEntropy::loss(Layer* layer, double** yTrue) {
-	double sum = 0;
+float BinaryCrossEntropy::loss(Layer* layer, float** yTrue) {
+	float sum = 0;
 	for (int i = 0; i < layer->batchSize; i++) {
 		for (int j = 0; j < layer->size; j++) {
 			sum += yTrue[i][j] * log(layer->neurons[i][j]) + (1 - yTrue[i][j]) * log(1 - layer->neurons[i][j]);
@@ -40,7 +40,7 @@ double BinaryCrossEntropy::loss(Layer* layer, double** yTrue) {
 	return -sum / layer->size;
 }
 
-void BinaryCrossEntropy::differentiate(Layer* layer, double** yTrue) {
+void BinaryCrossEntropy::differentiate(Layer* layer, float** yTrue) {
 	for (int i = 0; i < layer->batchSize; i++) {
 		for (int j = 0; j < layer->size; j++) {
 			layer->neuronGradient[i][j] = (-yTrue[i][j] / layer->neurons[i][j] + (1 - yTrue[i][j]) / (1 - layer->neurons[i][j])) / layer->size;
@@ -53,8 +53,8 @@ string BinaryCrossEntropy::toString() {
 	return toString;
 }
 
-double CategoricalCrossEntropy::loss(Layer* layer, double** yTrue) {
-	double sum = 0;
+float CategoricalCrossEntropy::loss(Layer* layer, float** yTrue) {
+	float sum = 0;
 	for (int i = 0; i < layer->batchSize; i++) {
 		for (int j = 0; j < layer->size; j++) {
 			if (yTrue[i][j] != 0) {
@@ -65,7 +65,7 @@ double CategoricalCrossEntropy::loss(Layer* layer, double** yTrue) {
 	return -sum / layer->size;
 }
 
-void CategoricalCrossEntropy::differentiate(Layer* layer, double** yTrue) {
+void CategoricalCrossEntropy::differentiate(Layer* layer, float** yTrue) {
 	for (int i = 0; i < layer->batchSize; i++) {
 		for (int j = 0; j < layer->size; j++) {
 			if (yTrue[i][j] != 0) {
@@ -83,7 +83,7 @@ string CategoricalCrossEntropy::toString() {
 	return toString;
 }
 
-double Accuracy::loss(Layer* layer, double** yTrue) {
+float Accuracy::loss(Layer* layer, float** yTrue) {
 	int sum = 0;
 	for (int i = 0; i < layer->batchSize; i++) {
 		int predMax = 0;
@@ -100,7 +100,7 @@ double Accuracy::loss(Layer* layer, double** yTrue) {
 	return sum;
 }
 
-void Accuracy::differentiate(Layer* layer, double** yTrue) {
+void Accuracy::differentiate(Layer* layer, float** yTrue) {
 	throw invalid_argument("Accuracy is a metric and cannot be used as a loss function");
 }
 
@@ -108,8 +108,8 @@ string Accuracy::toString() {
 	return "Accuracy";
 }
 
-double BinaryAccuracy::loss(Layer* layer, double** yTrue) {
-	double sum = 0;
+float BinaryAccuracy::loss(Layer* layer, float** yTrue) {
+	float sum = 0;
 	for (int i = 0; i < layer->batchSize; i++) {
 		for (int j = 0; j < layer->size; j++) {
 			if ((layer->neurons[i][j] >= 0.5 && yTrue[i][j] >= 0.5) || (layer->neurons[i][j] < 0.5 && yTrue[i][j] < 0.5)) {
@@ -120,7 +120,7 @@ double BinaryAccuracy::loss(Layer* layer, double** yTrue) {
 	return sum / layer->size;
 }
 
-void BinaryAccuracy::differentiate(Layer* layer, double** yTrue) {
+void BinaryAccuracy::differentiate(Layer* layer, float** yTrue) {
 	throw invalid_argument("BinaryAccuracy is a metric and cannot be used as a loss function");
 }
