+++ DenseNeuralNetwork/NeuralNetwork.h
@@ -14,14 +14,14 @@ public:
 	NeuralNetwork(string fileName);
 
 	void addLayer(Layer* layer);
-	void predict(double** input);
-	void forwardPropagate(double** input);
-	void backPropagate(Loss* lossFunction, double** yTrue);
-	void applyGradients(double learningRate);
-	void fit(Loss* lossFunction, double** X, double** y, double* losses, int numMetrics, Loss** metrics, TrainingParams* params);
-	void fit(Loss* lossFunction, int numData, double** X, double** y, int numMetrics, Loss** metrics, TrainingParams* params);
-	void test(Loss* lossFunction, int numData, double** X, double** y, int numMetrics, Loss** metrics);
-	void shuffle(int numData, double** X, double** y);
+	void predict(float** input);
+	void forwardPropagate(float** input);
+	void backPropagate(Loss* lossFunction, float** yTrue);
+	void applyGradients(float learningRate);
+	void fit(Loss* lossFunction, float** X, float** y, float* losses, int numMetrics, Loss** metrics, TrainingParams* params);
+	void fit(Loss* lossFunction, int numData, float** X, float** y, int numMetrics, Loss** metrics, TrainingParams* params);
+	void test(Loss* lossFunction, int numData, float** X, float** y, int numMetrics, Loss** metrics);
+	void shuffle(int numData, float** X, float** y);
 	void setTrainable(bool trainable);
 	void save(string fileName);
 	int getNumParameters();
